refactor(edda-serve): extract error, state, helpers, middleware modules
Split foundational types out of the 8.7K-line lib.rs into dedicated files:
- error.rs: AppError enum and IntoResponse impl
- state.rs: ServeConfig, AppState, ChronicleContext, PairingRequest
- helpers.rs: validate_iso8601, time_now_rfc3339
- middleware.rs: auth_middleware, is_localhost, generate_pairing_token

Part of GH-375.
